requirements/requirements.txt
python-decouple==3.3
requests==2.24.0
rich==5.2.0